Add support for RNA-seq pipeline with STAR 1-pass
Allow using the genomeDir files as the primary index,
skipping the 1st pass of STAR. Please, notice that no
novel junctions will be considered.

With this addition, all samples are independent of each
other, allowing their processing in parallel. To
complement this idea, an option to create separate json
files is provided